requirements.txt
altair==4.1.0
altgraph==0.17.2
argcomplete==1.12.3
argon2-cffi==21.1.0
astor==0.8.1
attrs==21.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.1
bleach==4.1.0
blinker==1.4
boto3==1.20.14
botocore==1.23.14
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==7.1.2
colorama==0.4.4
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
enum-compat==0.0.3
et-xmlfile==1.1.0
ezdxf==0.17.1
future==0.18.2
gitdb==4.0.9
GitPython==3.1.24
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
ipykernel==6.5.1
ipython==7.30.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
jmespath==0.10.0
jsonschema==4.2.1
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.0.0
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.1.6
numpy==1.21.4
openpyxl==3.0.9
packaging==21.3
pandas==1.1.0
pandocfilters==1.5.0
parso==0.8.2
pefile==2021.9.3
pickleshare==0.7.5
Pillow==8.4.0
prometheus-client==0.12.0
prompt-toolkit==3.0.23
protobuf==3.19.1
pyarrow==6.0.1
pycparser==2.21
pydeck==0.7.1
Pygments==2.10.0
pyinstaller==4.7
pyinstaller-hooks-contrib==2021.3
Pympler==0.9
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32==302
pywin32-ctypes==0.2.0
pywinpty==1.1.6
pyzmq==22.3.0
requests==2.26.0
s3transfer==0.5.0
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
streamlit==0.68.0
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
toolz==0.11.2
tornado==5.1.1
traitlets==5.1.1
typing_extensions==4.0.0
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
validators==0.18.2
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
XlsxWriter==3.0.2
zipp==3.6.0
